[REFACTOR] Organise react admin files
