+++ src/client/react/admin/components/AdminNavbar/SecondaryNavigation.js
@@ -4,12 +4,12 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import axios from 'axios';
-import API from '../../../API'
-import MenuButton from '../../../../../lib/react/components/MenuButton';
+import API from '../../../../API'
+import MenuButton from '../../../../../../lib/react/components/MenuButton';
 import SettingsMenu from './SettingsMenu';
 import { withRouter } from 'react-router-dom';
-import { logout } from '../../../actions/authActions';
-import LoginRefresher from '../../components/LoginRefresher';
+import { logout } from '../../../../actions/authActions';
+import LoginRefresher from '../../../components/LoginRefresher';
 
 
 const mapStateToProps = (state, ownProps) => {
