+++ src/client/react/admin/components/AdminNavbar/SettingsMenu.js
