+++ src/client/react/admin/components/AdminNavbar/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Navbar from '../../../../../lib/react/components/Navbar';
+import Navbar from '../../../../../../lib/react/components/Navbar';
 import MainNavigation from './MainNavigation';
 import SecondaryNavigation from './SecondaryNavigation';
 
-import '../../scss/admin/_navbar.scss';
+import '../../../scss/admin/_navbar.scss';
 
 
 class AdminNavbar extends React.Component {
