+++ src/client/react/admin/components/ViewError.js
