+++ src/client/react/admin/views/ChallengesView/ChallengeCard.js
