+++ src/client/react/admin/views/ChallengesView/ChallengeProfile.js
@@ -13,12 +13,12 @@ import {
 	setChallengeTitle,
 	setChallengeDescription,
 	setChallengeLocked
-} from '../../../graphql/challenge';
-import NotificationToaster from '../NotificationToaster';
-import MarkdownEditor from '../../../../../lib/react/components/MarkdownEditor';
+} from '../../../../graphql/challenge';
+import NotificationToaster from '../../components/NotificationToaster';
+import MarkdownEditor from '../../../../../../lib/react/components/MarkdownEditor';
 
-import '../../scss/components/_instruction-panel.scss';
-import '../../scss/admin/_markdown-preview.scss';
+import '../../../scss/components/_instruction-panel.scss';
+import '../../../scss/admin/_markdown-preview.scss';
 
 
 const QueryGetChallengeOptions = {
