+++ src/client/react/admin/views/GameStateView/gameStates.js
