+++ src/client/react/admin/views/InstructionArticlesView/InstructionArticleCard.js
