+++ src/client/react/admin/views/TeamsView/TeamCard.js
